+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * convert soft redirects to HTTP 308
 * update route handler params typing for Next
 * add `npm run typecheck` script
+* support `CRON_SECRET_NEXT` for rotation
 
 
 # 1.0.0 (2026-01-06)